version 1.0.1

+++ models/user.go
@@ -3,7 +3,7 @@ package models
 // 用户全信息，用于接口
 type UserInfo struct {
 	User_name  string  `json:"user_name"`
-	User_id    int32   `json:"user_id" gorm:"primaryKey"`
+	User_id    int64   `json:"user_id" gorm:"primaryKey"`
 	User_pwd   string  `json:"user_pwd"`
 	User_tag   string  `json:"user_tag"` // 兴趣标签的json字符串
 	Log        string  `json:"log"`      // 用户备注
@@ -18,7 +18,7 @@ type UserInfo struct {
 // 数据库user表信息，用于gorm
 type User struct {
 	User_name  string `json:"user_name"`
-	User_id    int32  `json:"user_id" gorm:"primaryKey"`
+	User_id    int64  `json:"user_id" gorm:"primaryKey"`
 	User_pwd   string `json:"user_pwd"`
 	User_tag   string `json:"user_tag"` // 兴趣标签的json字符串
 	Log        string `json:"log"`      // 用户备注
@@ -29,8 +29,8 @@ type User struct {
 
 // 数据库space表信息，用于gorm
 type Space struct {
-	User_id  int32
-	Query_id int32
+	User_id  int64
+	Query_id int64
 	Type     int16
 }
